requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.1
google-generativeai==0.3.2
pyyaml==6.0.1
websockets==12.0
pydantic==2.6.0
requests==2.31.0